requirements.txt
pymysql
pymongo
seaborn
plotly
fpdf
customtkinter